explicit data conversion to solve bbb building problems

+++ src/gmm_classifier.cpp
@@ -41,7 +41,7 @@ double gmm_classifier::NDgaussian(Behaviors::MODE mode, int n_cluster, VectorXd 
 
   vec = (features_vector.transpose() - Models[mode].mu.block(n_cluster,0,1,8)).transpose();
   std::cout<<"Test Mean diff"<<vec<<std::endl;
-  result = 1 / sqrt( pow(2 * 3.141516,8) * Models[mode].sigma_det[n_cluster]) * exp(-0.5 * vec.transpose() * Models[mode].sigma_inv[n_cluster] * vec);
+  result = 1 / sqrt( pow(2 * 3.141516,8) * Models[mode].sigma_det[n_cluster]) * exp((double)(-0.5 * vec.transpose() * Models[mode].sigma_inv[n_cluster] * vec));
 
   return result;
 }
